@@ -44,32 +44,28 @@
     "dist": "yarn compile && electron-builder --publish never"
   },
   "dependencies": {
-    "@electron/remote": "^2.0.4",
+    "@electron/remote": "^2.0.5",
     "@noble/ed25519": "^1.6.0",
     "@noble/secp256k1": "^1.5.5",
-    "@panter/vue-i18next": "^0.15.2",
+    "@vueform/multiselect": "^2.3.3",
+    "balm-ui": "^9.42.2",
     "bitcoinjs-lib": "^6.0.1",
-    "bitsharesjs": "^5.0.0",
-    "bootstrap-vue": "^2.0.3",
+    "bitsharesjs": "^6.0.0",
     "crypto-js": "^4.1.1",
     "dexie": "^3.2.1",
     "electron-devtools-installer": "^3.2.0",
-    "i18next": "^21.6.11",
-    "i18next-fs-backend": "^1.1.4",
+    "linebyline": "^1.3.0",
     "lzma": "^2.3.2",
-    "otpauth": "^7.0.10",
-    "readline": "^1.3.0",
+    "otpauth": "^7.0.11",
     "socket.io": "^4.4.1",
     "typeface-rajdhani": "1.1.13",
     "typeface-roboto": "1.1.13",
     "uuid": "^8.3.2",
-    "vue": "^2.6.10",
+    "vue": "^3.2.31",
+    "vue-i18n": "9",
     "vue-multiselect": "^2.1.6",
-    "vue-password-strength-meter": "^1.5.0",
-    "vue-router": "^3.1.3",
-    "vuex": "^3.1.1",
-    "ws": "^8.5.0",
-    "zxcvbn": "^4.4.2"
+    "vue-router": "^4.0.12",
+    "vuex": "^4.0.2"
   },
   "devDependencies": {
     "@babel/core": "^7.17.5",
@@ -79,17 +75,17 @@
     "@soda/friendly-errors-webpack-plugin": "^1.8.1",
     "babel-eslint": "^10.0.3",
     "babel-loader": "^8.2.3",
-    "css-loader": "^5.2.7",
-    "electron": "^17.0.1",
+    "css-loader": "^6.6.0",
+    "electron": "^17.1.0",
     "electron-builder": "^22.14.13",
     "eslint": "^8.9.0",
-    "eslint-plugin-vue": "^8.4.1",
+    "eslint-plugin-vue": "^8.5.0",
     "file-loader": "^6.2.0",
-    "sass": "^1.49.8",
+    "sass": "^1.49.9",
     "sass-loader": "^12.6.0",
     "source-map-support": "^0.5.21",
     "style-loader": "^3.3.1",
-    "vue-loader": "^15.9.8",
+    "vue-loader": "^17.0.0",
     "vue-template-compiler": "^2.6.14",
     "webpack": "^5.69.1",
     "webpack-cli": "^4.9.2",
