@@ -1,27 +1,29 @@
-import store from './store/index';
-import Vue from 'vue';
+import { createApp } from 'vue';
+import { createI18n } from 'vue-i18n';
+import VueRouter from 'vue-router';
 
-import BootstrapVue from 'bootstrap-vue';
+import BalmUI from 'balm-ui'; // Official Google Material Components
+import BalmUIPlus from 'balm-ui/dist/balm-ui-plus'; // BalmJS Team Material Components
+
+import router from './router/index.js';
+import store from './store/index';
+import BeetServer from './lib/BeetServer';
+import RendererLogger from './lib/RendererLogger';
+import fetchMessages from './lib/i18n';
 import Popups from './components/popups';
 import Beetframe from './components/beetframe';
-import VueRouter from 'vue-router';
-import router from './router/index.js';
-import i18next from 'i18next';
-import Backend from "i18next-fs-backend";
-import VueI18Next from '@panter/vue-i18next';
-import BeetServer from "./lib/BeetServer";
-import RendererLogger from "./lib/RendererLogger";
+
 import 'typeface-roboto';
 import 'typeface-rajdhani';
 
-import "./css/style.css";
+import 'bootstrap/dist/css/bootstrap.css'
+import 'bootstrap-vue-3/dist/bootstrap-vue-3.css'
+
+import './css/style.css';
 import './scss/beet.scss';
 
 const logger = new RendererLogger;
-Vue.config.errorHandler = function (err, vm, info) {
-  logger.error(err, vm, info);
-  console.log("error");
-};
+
 window.onerror = function (msg, url, lineNo, columnNo, error) {
   logger.error(error);
   console.log(error);
@@ -31,29 +33,36 @@ process.on('unhandledRejection', (reason, p) => {
   console.log('Unhandled Rejection at: Promise', p, 'reason:', reason);
   // application specific logging, throwing an error, or other logic here
 });
-Vue.use(VueI18Next);
-Vue.use(BootstrapVue);
-Vue.use(VueRouter);
-Vue.config.devtools = true;
-Vue.component('Popups', Popups);
-Vue.component('Beetframe', Beetframe);
+
 store.dispatch("SettingsStore/loadSettings");
 store.dispatch("WhitelistStore/loadWhitelist");
 
-i18next.use(Backend).init({
-  lng: store.state.SettingsStore.settings.locale.iso,
-  ns: ['common', 'operations'],
-  defaultNS: 'common',
-  backend: {
-    loadPath: `${__dirname}/translations/{{ns}}/{{lng}}.json`,
-  }
-});
-const i18n = new VueI18Next(i18next);
+const messages = fetchMessages();
 
-const app = new Vue({
-  router,
-  store,
-  i18n: i18n
-}).$mount('#app');
+const i18n = createI18n({
+  locale: store.state.SettingsStore.settings.locale.iso,
+  fallbackLocale: 'en',
+  messages
+})
+
+const app = createApp({});
+
+app.config.devtools = true;
+app.config.errorHandler = function (err, vm, info) {
+  logger.error(err, vm, info);
+  console.log("error");
+};
+
+app.component('Popups', Popups);
+app.component('Beetframe', Beetframe);
+
+app.use(i18n);
+app.use(VueRouter);
+app.use(BalmUI);
+app.use(BalmUIPlus);
+
+app.use(router);
+app.use(store);
+app.mount('#app');
 
 BeetServer.initialize(app, 60555);
