@@ -1,13 +1,34 @@
+<script setup>
+    import getBlockchain from "../lib/blockchains/blockchainFactory";
+    import {formatChain} from "../lib/formatter";
+    import { shell } from 'electron';
+
+    function getChainLabel(chain) {
+        return formatChain(chain);
+    }
+    function getExplorer(account) {
+        return getBlockchain(account.chain).getExplorer(account);
+    }
+    function getAccessType(chain) {
+        return getBlockchain(chain).getAccessType();
+    }
+    function openExplorer(account) {
+        shell.openExternal(this.getExplorer(account));
+    }
+
+    const account = defineProps(["account"]);
+</script>
+
 <template>
     <div class="account-details mt-3">
         <p class="mb-1 font-weight-bold small">
-            {{ $t('account_details_lbl') }}
+            {{ $t('common.account_details_lbl') }}
         </p>
         <table class="table small table-striped table-sm">
             <tbody v-if="account">
                 <tr>
                     <td class="text-left">
-                        {{ $t('account_details_chaim_lbl') }}
+                        {{ $t('common.account_details_chaim_lbl') }}
                     </td>
                     <td class="text-right">
                         {{ getChainLabel(account.chain) }}
@@ -15,7 +36,7 @@
                 </tr>
                 <tr>
                     <td class="text-left">
-                        {{ getAccessType(account.chain) == "account" ? $t('account_details_name_lbl') : $t('account_details_address_lbl') }}
+                        {{ getAccessType(account.chain) == "account" ? $t('common.account_details_name_lbl') : $t('common.account_details_address_lbl') }}
                     </td>
                     <td class="text-right">
                         {{ account.accountName }}
@@ -23,7 +44,7 @@
                 </tr>
                 <tr v-if="account.accountName != account.accountID">
                     <td class="text-left">
-                        {{ $t('account_details_id_lbl') }}
+                        {{ $t('common.account_details_id_lbl') }}
                     </td>
                     <td class="text-right">
                         {{ account.accountID }}
@@ -31,13 +52,10 @@
                 </tr>
                 <tr v-if="getExplorer(account)">
                     <td class="text-left">
-                        {{ $t('account_details_explorer_lbl') }}
+                        {{ $t('common.account_details_explorer_lbl') }}
                     </td>
                     <td class="text-right">
-                        <a
-                            href="#"
-                            @click="openExplorer(account)"
-                        >
+                        <a href="#" @click="openExplorer(account)">
                             Click here
                         </a>
                     </td>
@@ -47,28 +65,3 @@
         </table>
     </div>
 </template>
-<script>
-    import getBlockchain from "../lib/blockchains/blockchainFactory";
-    import {formatChain} from "../lib/formatter";    
-    import { shell } from 'electron';
-
-    export default {
-        name: "AccountDetails",
-        i18nOptions: { namespaces: "common" },
-        props: ["account"],
-        methods: {
-            getChainLabel: function(chain) {
-                return formatChain(chain);
-            },
-            getExplorer: function(account) {
-                return getBlockchain(account.chain).getExplorer(account);
-            },
-            getAccessType: function(chain) {
-                return getBlockchain(chain).getAccessType();
-            },
-            openExplorer: function(account) {
-                shell.openExternal(this.getExplorer(account));
-            }
-        }
-    };
-</script>
