@@ -3,39 +3,39 @@
         id="step2"
     >
         <h4 class="h4 mt-3 font-weight-bold">
-            {{ $t('step_counter',{ 'step_no' : 2}) }}
+            {{ $t('common.step_counter',{ 'step_no' : 2}) }}
         </h4>
         <p
             class="mb-2 font-weight-bold"
         >
-            {{ $t('account_name', { 'chain' : selectedChain}) }}
+            {{ $t('common.account_name', { 'chain' : selectedChain}) }}
         </p>
         <input
             id="inputAccount"
             v-model="accountname"
             type="text"
             class="form-control mb-3"
-            :placeholder="$t('account_name', { 'chain' : selectedChain})"
+            :placeholder="$t('common.account_name', { 'chain' : selectedChain})"
             required
         >
         <p class="my-3 font-weight-normal">
-            {{ $t('keys_cta') }}
+            {{ $t('common.keys_cta') }}
         </p>
         <template v-if="requiredFields.memo !== null">
             <p class="mb-2 font-weight-bold">
-                {{ $t('memo_authority') }}
+                {{ $t('common.memo_authority') }}
             </p>
             <input
                 id="inputMemo"
                 v-model="memopk"
                 type="password"
                 class="form-control mb-3 small"
-                :placeholder="$t('memo_authority_placeholder')"
+                :placeholder="$t('common.memo_authority_placeholder')"
                 required
             >
         </template>
         <p class="my-3 font-weight-normal">
-            {{ $t('use_only_for_messages_and_proof') }}
+            {{ $t('common.use_only_for_messages_and_proof') }}
         </p>
     </div>
 
@@ -46,7 +46,6 @@
 
     export default {
         name: "ImportMemo",
-        i18nOptions: { namespaces: "common" },
         props: [ "selectedChain" ],
         data() {
             return {
