@@ -1,41 +1,31 @@
-<template>
-    <div class="top">
-        <div class="container-fluid">
-            <div class="row">
-                <div class="col-12 largeHeader">
-                    <img
-                        src="img/beet.png"
-                        alt
-                        width="150"
-                        height="150"
-                    >
-                    <div class="beet-typo mb-3">
-                        beet
-                    </div>
-                    <h3 class="h3 mb-3 font-weight-normal">
-                        {{ $t("welcome" ) }}
-                    </h3>
-                    <LangSelect ref="langswitch" />
-                </div>
-            </div>
-        </div>
-    </div>
-</template>
- 
-<script>
+<script setup>
     import LangSelect from "./lang-select";
     import RendererLogger from "../lib/RendererLogger";
     const logger = new RendererLogger();
 
-    export default {
-        name: "HeaderGuest",
-        i18nOptions: { namespaces: "common" },
-        components: { LangSelect },
-        data() {
-            return {};
-        },
-        mounted() {
-            logger.debug("Standard Header mounted");
-        }
-    };
+    onMounted(() => {
+      logger.debug("Standard Header mounted");
+    });
 </script>
+
+<template>
+    <div class="top">
+      <ui-grid class="container-fluid">
+          <ui-grid-cell class="largeHeader" columns="12">
+                <img
+                    src="img/beet.png"
+                    alt
+                    width="150"
+                    height="150"
+                >
+                <div class="beet-typo">
+                    beet
+                </div>
+                <h3 class="h3">
+                    {{ $t("common.welcome" ) }}
+                </h3>
+                <LangSelect ref="langswitch" />
+            </ui-grid-cell>
+        </ui-grid>
+    </div>
+</template>
