@@ -8,17 +8,17 @@
         no-close-on-backdrop
         hide-header-close
         hide-footer
-        :title="$t('operations:account_id.title')"
+        :title="$t('operations.account_id.title')"
     >
         <div
             v-b-tooltip.hover
-            :title="$t('operations:link.request_tooltip')"
+            :title="$t('operations.link.request_tooltip')"
         >
-            {{ $t('operations:link.request', {appName: incoming.appName, origin: incoming.origin, chain: incoming.chain }) }} &#10068;
+            {{ $t('operations.link.request', {appName: incoming.appName, origin: incoming.origin, chain: incoming.chain }) }} &#10068;
         </div>
         <br>
         <div v-if="existingLinks.length>0">
-            {{ $t('operations:link.request_fresh', {chain: incoming.chain }) }}
+            {{ $t('operations.link.request_fresh', {chain: incoming.chain }) }}
         </div>
         <br>
         <AccountSelect
@@ -26,7 +26,7 @@
             v-model="chosenAccount"
             :chain="incoming.chain"
             :existing="existingLinks"
-            :cta="$t('operations:link.request_cta')"
+            :cta="$t('operations.link.request_cta')"
             extraclass="accountProvide"
         />
         <b-btn
@@ -35,7 +35,7 @@
             block
             @click="clickedAllow"
         >
-            {{ $t('operations:link.accept_btn') }}
+            {{ $t('operations.link.accept_btn') }}
         </b-btn>
         <b-btn
             class="mt-1"
@@ -43,7 +43,7 @@
             block
             @click="_clickedDeny"
         >
-            {{ $t('operations:link.reject_btn') }}
+            {{ $t('operations.link.reject_btn') }}
         </b-btn>
     </b-modal>
 </template>
@@ -82,7 +82,7 @@
         },
         methods: {
             _onShow: function () {
-                this.error=false;                
+                this.error=false;
                 this.chosenAccount={trackId: 0};
             },
             _execute: function () {
@@ -102,4 +102,4 @@
             }
         }
     };
-</script> 
+</script>
