@@ -7,11 +7,11 @@
         no-close-on-backdrop
         hide-header-close
         hide-footer
-        :title="$t('operations:rawsig.title')"
+        :title="$t('operations.rawsig.title')"
     >
         <table
             v-b-tooltip.hover
-            :title="$t('operations:rawsig.request',
+            :title="$t('operations.rawsig.request',
                 { appName: incoming.appName,
                   origin: incoming.origin,
                   chain: getChainLabel(incoming.chain),
@@ -50,13 +50,13 @@
                         Action
                     </td>
                     <td class="text-right">
-                        {{getRequestType() ? $t('operations:rawsig.sign_btn') : $t('operations:rawsig.sign_and_broadcast_btn') }}
+                        {{getRequestType() ? $t('operations.rawsig.sign_btn') : $t('operations.rawsig.sign_and_broadcast_btn') }}
                     </td>
                 </tr>
             </tbody>
         </table>
         <p class="mb-1 font-weight-bold small">
-            {{ $t('operations:general.content') }}
+            {{ $t('operations.general.content') }}
         </p>
         <div
             v-if="!!incoming.visualized"
@@ -73,14 +73,14 @@
 }
 </code>
         </pre>
-        {{ $t('operations:rawsig.request_cta') }}
+        {{ $t('operations.rawsig.request_cta') }}
         <b-btn
             class="mt-3"
             variant="success"
             block
             @click="_clickedAllow"
         >
-            {{ getRequestType() == "sign" ? $t('operations:rawsig.sign_btn') : $t('operations:rawsig.sign_and_broadcast_btn') }}
+            {{ getRequestType() == "sign" ? $t('operations.rawsig.sign_btn') : $t('operations.rawsig.sign_and_broadcast_btn') }}
         </b-btn>
         <b-btn
             class="mt-1"
@@ -88,7 +88,7 @@
             block
             @click="_clickedDeny"
         >
-            {{ $t('operations:rawsig.reject_btn') }}
+            {{ $t('operations.rawsig.reject_btn') }}
         </b-btn>
     </b-modal>
 </template>
