@@ -1,15 +1,11 @@
-import Vue from 'vue';
-import Vuex from 'vuex';
+import { createStore } from 'vuex';
 import WalletStore from './modules/WalletStore.js';
 import WhitelistStore from './modules/WhitelistStore.js';
 import OriginStore from './modules/OriginStore.js';
 import SettingsStore from './modules/SettingsStore.js';
 import AccountStore from './modules/AccountStore.js';
 
-Vue.config.devtools = true;
-Vue.use(Vuex);
-
-const store = new Vuex.Store({
+export default createStore({
     modules: {
         WalletStore,
         WhitelistStore,
@@ -18,5 +14,3 @@ const store = new Vuex.Store({
         AccountStore
     }
 });
-
-export default store;
